Initial version
The minute weather dataset comes from the same source as the daily weather dataset that was used for the hands-on activities in the previous modules. The difference is that the minute weather dataset contains raw sensor measurements captured at one-minute intervals, not processed like the daily weather dataset. The data is in the file minute_weather.csv, which is a comma-separated file.

As with the daily weather data, this data comes from a weather station located in San Diego, California. The weather station is equipped with sensors that capture weather-related measurements such as air temperature, air pressure, and relative humidity. Data was collected for a period of three years, from September 2011 to September 2014, to ensure that sufficient data for different seasons and weather conditions is captured.

Each row in minute_weather.csv contains weather data captured for a one-minute interval.